refactor(gamedb): split 4391-line monolith into focused modules under src/commands/gamedb/
- gamedb-utils.ts: shared helpers (search encoding, permissions, autocomplete, etc.)
- gamedb-profile.service.ts: profile builder, action rows, text formatting
- gamedb-csv-import.service.ts: CSV scoring, platform mapping, IGDB select options
- gamedb-add.command.ts: GameDbAddCommand (add, refresh-release-info slashes)
- gamedb-csv-import.command.ts: GameDbCsvImportCommand (csv-import slash + all CSV handlers)
- gamedb-view.command.ts: GameDbViewCommand (view slash + gamedb-action button dispatcher)
- gamedb-search.command.ts: GameDbSearchCommand (search slash + pagination handlers)
- gamedb-completion.command.ts: GameDbCompletionCommand (completion wizard + nowplaying modal)
- gamedb-thread.command.ts: GameDbThreadCommand (now playing thread modal handler)
- gamedb.command.ts becomes a barrel re-exporting all command classes
- Removes unreachable legacy text-based completion wizard (runCompletionWizard)
- Updates nominate.command.ts and GameReleaseAnnouncementService.ts to use direct imports
