+++ src/commands/gamedb/gamedb-profile.service.ts
@@ -1,0 +1,735 @@
+import {
+  ActionRowBuilder,
+  AttachmentBuilder,
+  ButtonBuilder,
+  ButtonStyle,
+  MessageFlags,
+  type ButtonInteraction,
+  type CommandInteraction,
+  type ModalSubmitInteraction,
+  type StringSelectMenuInteraction,
+} from "discord.js";
+import {
+  ButtonBuilder as V2ButtonBuilder,
+  ContainerBuilder,
+  SectionBuilder,
+  SeparatorBuilder,
+  TextDisplayBuilder,
+  ThumbnailBuilder,
+} from "@discordjs/builders";
+import { SeparatorSpacingSize } from "discord-api-types/v10";
+import { safeReply } from "../../functions/InteractionUtils.js";
+import { formatPlatformDisplayName } from "../../functions/PlatformDisplay.js";
+import { formatTableDate } from "../profile.command.js";
+import { getHltbCacheByGameId } from "../../classes/HltbCache.js";
+import { getThreadsByGameId } from "../../classes/Thread.js";
+import { renderUsernameWithEmoji } from "../../services/UserEmojiService.js";
+import { padCommandName } from "../help.command.js";
+import Game, { type GameSource, type IGame, type IRelease } from "../../classes/Game.js";
+import {
+  buildComponentsV2Flags,
+  getSearchRowsFromComponents,
+  isHltbImportEligible,
+} from "./gamedb-utils.js";
+
+export type GameProfileRenderContext = {
+  guildId?: string;
+  includeInlineButtons?: boolean;
+  prefaceText?: string;
+};
+
+type GameProfileResult = {
+  components: Array<ContainerBuilder | ActionRowBuilder<ButtonBuilder>>;
+  files: AttachmentBuilder[];
+  hasThread: boolean;
+  featuredVideoUrl: string | null;
+  canMarkThumbnailBad: boolean;
+  isThumbnailBad: boolean;
+  isThumbnailApproved: boolean;
+  isReleased: boolean;
+};
+
+export function chunkText(text: string, size: number): string[] {
+  if (!text) return ["No description available."];
+  const chunks: string[] = [];
+  for (let i = 0; i < text.length; i += size) {
+    chunks.push(text.slice(i, i + size));
+  }
+  return chunks;
+}
+
+export function formatQuotedDescription(description: string): string {
+  if (!description.trim()) {
+    return "> No description available.";
+  }
+  return description
+    .split(/\r?\n/)
+    .map((line) => (line.trim().length ? `> ${line}` : ""))
+    .join("\n");
+}
+
+export function buildListFieldValue(lines: string[], maxLength: number): string {
+  if (!lines.length) return "None";
+  const output: string[] = [];
+  let currentLength = 0;
+
+  for (let i = 0; i < lines.length; i += 1) {
+    const line = lines[i];
+    const nextLength = currentLength + line.length + 1;
+    if (nextLength > maxLength) {
+      const remaining = lines.length - i;
+      output.push(`…and ${remaining} more`);
+      break;
+    }
+    output.push(line);
+    currentLength = nextLength;
+  }
+
+  return output.join("\n");
+}
+
+export function trimTextDisplayContent(content: string): string {
+  if (content.length <= 4000) {
+    return content;
+  }
+  return `${content.slice(0, 3997)}...`;
+}
+
+export function isGameReleased(game: IGame, releases: IRelease[]): boolean {
+  const releaseDates: Date[] = [];
+  for (const release of releases) {
+    if (release.releaseDate instanceof Date && !Number.isNaN(release.releaseDate.getTime())) {
+      releaseDates.push(release.releaseDate);
+    }
+  }
+  if (
+    game.initialReleaseDate instanceof Date &&
+    !Number.isNaN(game.initialReleaseDate.getTime())
+  ) {
+    releaseDates.push(game.initialReleaseDate);
+  }
+
+  if (!releaseDates.length) {
+    return true;
+  }
+
+  const earliestReleaseTimestamp = releaseDates
+    .map((releaseDate) => releaseDate.getTime())
+    .reduce((earliest, current) => (current < earliest ? current : earliest));
+  return earliestReleaseTimestamp <= Date.now();
+}
+
+export function buildGameProfileActionRow(
+  gameId: number,
+  hasThread: boolean,
+  featuredVideoUrl: string | null,
+  canMarkThumbnailBad: boolean,
+  isThumbnailBad: boolean,
+  isThumbnailApproved: boolean,
+  isReleased: boolean,
+  disableVideo = false,
+): ActionRowBuilder<ButtonBuilder>[] {
+  const addNowPlaying = new ButtonBuilder()
+    .setCustomId(`gamedb-action:nowplaying:${gameId}`)
+    .setLabel("Add to Now Playing List")
+    .setStyle(ButtonStyle.Primary);
+  const viewFeaturedVideo = new ButtonBuilder()
+    .setCustomId(`gamedb-action:video:${gameId}`)
+    .setLabel("View Featured Video")
+    .setStyle(ButtonStyle.Secondary)
+    .setDisabled(disableVideo);
+  const rows: ActionRowBuilder<ButtonBuilder>[] = [];
+  const primaryRow = new ActionRowBuilder<ButtonBuilder>().addComponents(addNowPlaying);
+  if (isReleased) {
+    const addCompletion = new ButtonBuilder()
+      .setCustomId(`gamedb-action:completion:${gameId}`)
+      .setLabel("Add Completion")
+      .setStyle(ButtonStyle.Primary);
+    primaryRow.addComponents(addCompletion);
+  }
+  if (!hasThread) {
+    const addThread = new ButtonBuilder()
+      .setCustomId(`gamedb-action:thread:${gameId}`)
+      .setLabel("Create Now Playing Thread")
+      .setStyle(ButtonStyle.Primary);
+    primaryRow.addComponents(addThread);
+  }
+  if (featuredVideoUrl) {
+    primaryRow.addComponents(viewFeaturedVideo);
+  }
+  rows.push(primaryRow);
+  if (canMarkThumbnailBad && !isThumbnailBad && !isThumbnailApproved) {
+    const badThumbnail = new ButtonBuilder()
+      .setCustomId(`gamedb-action:bad-thumb:${gameId}`)
+      .setLabel("Bad Thumbnail")
+      .setStyle(ButtonStyle.Danger);
+    const goodThumbnail = new ButtonBuilder()
+      .setCustomId(`gamedb-action:good-thumb:${gameId}`)
+      .setLabel("Good Thumbnail")
+      .setStyle(ButtonStyle.Secondary);
+    const thumbRow = new ActionRowBuilder<ButtonBuilder>().addComponents(
+      badThumbnail,
+      goodThumbnail,
+    );
+    rows.push(thumbRow);
+  }
+  return rows;
+}
+
+export async function buildGameProfile(
+  gameId: number,
+  interaction?:
+    | CommandInteraction
+    | StringSelectMenuInteraction
+    | ButtonInteraction
+    | ModalSubmitInteraction
+    | GameProfileRenderContext,
+  source?: GameSource,
+): Promise<GameProfileResult | null> {
+  try {
+    const game = await Game.getGameById(gameId, source);
+    if (!game) {
+      return null;
+    }
+
+    const releases = await Game.getGameReleases(gameId);
+    const platforms = await Game.getAllPlatforms();
+    const regions = await Game.getAllRegions();
+    const associations = await Game.getGameAssociations(gameId);
+    const nowPlayingMembers = await Game.getNowPlayingMembers(gameId);
+    const collectionOwners = await Game.getGameCollectionOwners(gameId);
+    const completions = await Game.getGameCompletions(gameId);
+    const alternateVersions = await Game.getAlternateVersions(gameId);
+    const linkedThreads = await getThreadsByGameId(gameId);
+
+    const platformMap = new Map(platforms.map((p) => [p.id, p.name]));
+    const regionMap = new Map(regions.map((r) => [r.id, r.name]));
+
+    const description = game.description || "No description available.";
+    const container = new ContainerBuilder();
+    const includeInlineButtons =
+      (interaction as GameProfileRenderContext | undefined)?.includeInlineButtons ?? true;
+    const prefaceText = (interaction as GameProfileRenderContext | undefined)?.prefaceText?.trim();
+
+    const files: AttachmentBuilder[] = [];
+    const isThumbnailBad = Boolean(game.thumbnailBad);
+    const isThumbnailApproved = Boolean(game.thumbnailApproved);
+    const isReleased = isGameReleased(game, releases);
+    const primaryArt = isThumbnailBad ? game.imageData : (game.artData ?? game.imageData);
+    if (primaryArt) {
+      files.push(new AttachmentBuilder(primaryArt, { name: "game_image.png" }));
+    }
+
+    const rpgClubSections: string[] = [];
+    const pushRpgClubSection = (title: string, value: string | null): void => {
+      if (!value) return;
+      rpgClubSections.push(`**${title}**\n> ${value}`);
+    };
+
+    const gotmNomineesByRound = new Map<number, string[]>();
+    associations.gotmNominations.forEach((nom) => {
+      const list = gotmNomineesByRound.get(nom.round) ?? [];
+      list.push(renderUsernameWithEmoji(nom.userId, nom.username));
+      gotmNomineesByRound.set(nom.round, list);
+    });
+    const nrGotmNomineesByRound = new Map<number, string[]>();
+    associations.nrGotmNominations.forEach((nom) => {
+      const list = nrGotmNomineesByRound.get(nom.round) ?? [];
+      list.push(renderUsernameWithEmoji(nom.userId, nom.username));
+      nrGotmNomineesByRound.set(nom.round, list);
+    });
+
+    if (associations.gotmWins.length) {
+      const lines = associations.gotmWins.map((win) => {
+        const nominees = gotmNomineesByRound.get(win.round) ?? [];
+        if (!nominees.length) {
+          return `Round ${win.round}`;
+        }
+        return `Round ${win.round} (nominated by ${nominees.join(",  ")})`;
+      });
+      pushRpgClubSection("GOTM Round(s)", lines.join("\n"));
+    }
+
+    if (associations.nrGotmWins.length) {
+      const lines = associations.nrGotmWins.map((win) => {
+        const nominees = nrGotmNomineesByRound.get(win.round) ?? [];
+        if (!nominees.length) {
+          return `Round ${win.round}`;
+        }
+        return `Round ${win.round} (nominated by ${nominees.join(",  ")})`;
+      });
+      pushRpgClubSection("NR-GOTM Round(s)", lines.join("\n"));
+    }
+
+    const threadId =
+      associations.gotmWins.find((w) => w.threadId)?.threadId ??
+      associations.nrGotmWins.find((w) => w.threadId)?.threadId ??
+      nowPlayingMembers.find((p) => p.threadId)?.threadId ??
+      linkedThreads[0] ??
+      null;
+    const headerLink = threadId
+      ? `https://discord.com/channels/${interaction?.guildId ?? "@me"}/${threadId}`
+      : null;
+    const headerLines = [
+      `## ${headerLink ? `[${game.title}](${headerLink})` : game.title}`,
+    ];
+
+    const redditUrlRaw =
+      associations.gotmWins.find((w) => w.redditUrl)?.redditUrl ??
+      associations.nrGotmWins.find((w) => w.redditUrl)?.redditUrl ??
+      null;
+    const redditUrl = redditUrlRaw === "__NO_VALUE__" ? null : redditUrlRaw;
+    if (redditUrl) {
+      pushRpgClubSection("Reddit Discussion Thread", `[Reddit Link](${redditUrl})`);
+    }
+
+    if (nowPlayingMembers.length) {
+      const MAX_NOW_PLAYING_DISPLAY = 12;
+      const lines = nowPlayingMembers.slice(0, MAX_NOW_PLAYING_DISPLAY).map((member) => {
+        const name = member.globalName ?? member.username ?? member.userId;
+        return renderUsernameWithEmoji(member.userId, name);
+      });
+
+      if (nowPlayingMembers.length > MAX_NOW_PLAYING_DISPLAY) {
+        const remaining = nowPlayingMembers.length - MAX_NOW_PLAYING_DISPLAY;
+        lines.push(`…and ${remaining} more playing now.`);
+      }
+
+      pushRpgClubSection("Now Playing", lines.join(",  "));
+    }
+
+    if (collectionOwners.length) {
+      const MAX_OWNERS_DISPLAY = 12;
+      const lines = collectionOwners.slice(0, MAX_OWNERS_DISPLAY).map((member) => {
+        const name = member.globalName ?? member.username ?? member.userId;
+        return renderUsernameWithEmoji(member.userId, name);
+      });
+
+      if (collectionOwners.length > MAX_OWNERS_DISPLAY) {
+        const remaining = collectionOwners.length - MAX_OWNERS_DISPLAY;
+        lines.push(`…and ${remaining} more own this.`);
+      }
+
+      pushRpgClubSection("Owned By", lines.join(",  "));
+    }
+
+    if (completions.length) {
+      const MAX_COMPLETIONS_DISPLAY = 12;
+      const uniqueCompletions = new Map<string, (typeof completions)[number]>();
+      completions.forEach((member) => {
+        if (!uniqueCompletions.has(member.userId)) {
+          uniqueCompletions.set(member.userId, member);
+        }
+      });
+      const uniqueList = Array.from(uniqueCompletions.values());
+      const lines = uniqueList.slice(0, MAX_COMPLETIONS_DISPLAY).map((member) => {
+        const name = member.globalName ?? member.username ?? member.userId;
+        return renderUsernameWithEmoji(member.userId, name);
+      });
+
+      if (uniqueList.length > MAX_COMPLETIONS_DISPLAY) {
+        const remaining = uniqueList.length - MAX_COMPLETIONS_DISPLAY;
+        lines.push(`…and ${remaining} more completed this.`);
+      }
+
+      pushRpgClubSection("Completed By", lines.join(",  "));
+    }
+
+    const gotmWinRounds = new Set(associations.gotmWins.map((win) => win.round));
+    const nrGotmWinRounds = new Set(associations.nrGotmWins.map((win) => win.round));
+
+    const gotmNominations = associations.gotmNominations.filter(
+      (nom) => !gotmWinRounds.has(nom.round),
+    );
+    if (gotmNominations.length) {
+      const lines = gotmNominations.map(
+        (nom) => `Round ${nom.round} - ${renderUsernameWithEmoji(nom.userId, nom.username)}`,
+      );
+      pushRpgClubSection("GOTM Nominations", lines.join(",  "));
+    }
+
+    const nrGotmNominations = associations.nrGotmNominations.filter(
+      (nom) => !nrGotmWinRounds.has(nom.round),
+    );
+    if (nrGotmNominations.length) {
+      const lines = nrGotmNominations.map(
+        (nom) => `Round ${nom.round} - ${renderUsernameWithEmoji(nom.userId, nom.username)}`,
+      );
+      pushRpgClubSection("NR-GOTM Nominations", lines.join(",  "));
+    }
+
+    const bodyParts: Array<{ content: string; accessory?: V2ButtonBuilder }> = [];
+    bodyParts.push({ content: `**Description**\n${formatQuotedDescription(description)}` });
+
+    const labelCandidates: string[] = [];
+    let releasesByDate: Map<string, string[]> | null = null;
+
+    if (releases.length > 0) {
+      const sortedReleases = [...releases].sort((a, b) => {
+        const aTime = a.releaseDate ? a.releaseDate.getTime() : Number.POSITIVE_INFINITY;
+        const bTime = b.releaseDate ? b.releaseDate.getTime() : Number.POSITIVE_INFINITY;
+        return aTime - bTime;
+      });
+
+      const releaseMap = new Map<string, string[]>();
+      sortedReleases.forEach((r) => {
+        const platformName = (formatPlatformDisplayName(platformMap.get(r.platformId))
+          ?? "Unknown Platform").trim();
+        const regionName = (regionMap.get(r.regionId) || "Unknown Region").trim();
+        const regionSuffix = regionName === "Worldwide" ? "" : ` (${regionName})`;
+        const releaseDate = r.releaseDate ? formatTableDate(r.releaseDate) : "TBD";
+        const format = r.format ? `(${r.format}) ` : "";
+        const platformLabel = `${platformName} ${regionSuffix}${format}`.trim();
+        const list = releaseMap.get(releaseDate) ?? [];
+        list.push(platformLabel);
+        releaseMap.set(releaseDate, list);
+      });
+      releasesByDate = releaseMap;
+    }
+
+    const hltbCache = await getHltbCacheByGameId(gameId);
+    const canImportHltb = isHltbImportEligible(game, Boolean(hltbCache));
+    const hltbLabels: string[] = [];
+
+    if (releasesByDate) {
+      labelCandidates.push(...releasesByDate.keys());
+    }
+
+    if (hltbCache) {
+      if (hltbCache.main) hltbLabels.push("Main");
+      if (hltbCache.mainSides) hltbLabels.push("Main + Sides");
+      if (hltbCache.completionist) hltbLabels.push("Completionist");
+      if (hltbCache.singlePlayer) hltbLabels.push("Single-Player");
+      if (hltbCache.coOp) hltbLabels.push("Co-Op");
+      if (hltbCache.vs) hltbLabels.push("Vs.");
+    }
+
+    labelCandidates.push(...hltbLabels);
+    const padWidth = labelCandidates.reduce((max, label) => Math.max(max, label.length), 0) + 1;
+
+    if (releasesByDate) {
+      const releaseField = Array.from(releasesByDate.entries())
+        .map(([dateLabel, platformsForDate]) =>
+          `\n> **\`\` ${padCommandName(dateLabel, padWidth)}\`\`**  ` +
+          platformsForDate.join(", "),
+        )
+        .join("");
+      bodyParts.push({ content: `**Releases** ${releaseField}` });
+    }
+
+    if (hltbCache) {
+      const hltbLines: string[] = [];
+      if (hltbCache.main)
+        hltbLines.push(
+          `> **\`\` ${padCommandName("Main", padWidth)}\`\`**  ${hltbCache.main}`,
+        );
+      if (hltbCache.mainSides)
+        hltbLines.push(
+          `> **\`\` ${padCommandName("Main + Sides", padWidth)}\`\`**  ${hltbCache.mainSides}`,
+        );
+      if (hltbCache.completionist)
+        hltbLines.push(
+          `> **\`\` ${padCommandName("Completionist", padWidth)}\`\`**  ${hltbCache.completionist}`,
+        );
+      if (hltbCache.singlePlayer)
+        hltbLines.push(
+          `> **\`\` ${padCommandName("Single-Player", padWidth)}\`\`**  ${hltbCache.singlePlayer}`,
+        );
+      if (hltbCache.coOp)
+        hltbLines.push(
+          `> **\`\` ${padCommandName("Co-Op", padWidth)}\`\`**  ${hltbCache.coOp}`,
+        );
+      if (hltbCache.vs)
+        hltbLines.push(
+          `> **\`\` ${padCommandName("Vs.", padWidth)}\`\`**  ${hltbCache.vs}`,
+        );
+      if (hltbLines.length) {
+        bodyParts.push({ content: `**HowLongToBeat™**\n${hltbLines.join("\n")}` });
+      }
+    } else if (canImportHltb && includeInlineButtons) {
+      const importHltb = new V2ButtonBuilder()
+        .setCustomId(`gamedb-action:hltb-import:${gameId}`)
+        .setLabel("Import HLTB Data")
+        .setStyle(ButtonStyle.Secondary);
+      bodyParts.push({
+        content: "**HowLongToBeat™**\n> No HLTB data cached.",
+        accessory: importHltb,
+      });
+    }
+
+    const series = await Game.getGameSeries(gameId);
+    const detailSections: string[] = [];
+
+    if (series) {
+      detailSections.push(`**Series / Collection**\n> ${series}`);
+    }
+
+    if (alternateVersions.length) {
+      const lines = alternateVersions.map(
+        (alt) => `> **${alt.title}**`,
+      );
+      const value = buildListFieldValue(lines, 2000);
+      detailSections.push(`**Alternate Versions**\n${value}`);
+    }
+
+    if (detailSections.length) {
+      bodyParts.push({ content: detailSections.join("\n\n") });
+    }
+
+    if (rpgClubSections.length) {
+      bodyParts.push({ content: rpgClubSections.join("\n\n") });
+    }
+
+    const igdbIdText = game.igdbId ? String(game.igdbId) : "N/A";
+    const sourceLabel = source === "API" ? "API" : "OracleSQL";
+    bodyParts.push({
+      content: `-# GameDB ID: ${game.id} | IGDB ID: ${igdbIdText} | Source: ${sourceLabel}`,
+    });
+
+    const headerBlock = trimTextDisplayContent(headerLines.join("\n"));
+    const bodyBlocks = bodyParts
+      .map((block) => ({
+        content: trimTextDisplayContent(block.content),
+        accessory: block.accessory,
+      }))
+      .filter((block) => block.content.length > 0);
+
+    if (prefaceText) {
+      container.addTextDisplayComponents(
+        new TextDisplayBuilder().setContent(trimTextDisplayContent(prefaceText)),
+      );
+      container.addSeparatorComponents(
+        new SeparatorBuilder().setSpacing(SeparatorSpacingSize.Large).setDivider(true),
+      );
+    }
+
+    const thumbnailUrl = primaryArt ? "attachment://game_image.png" : (game.coverUrl ?? null);
+
+    if (thumbnailUrl) {
+      if (headerBlock.length > 0) {
+        container.addTextDisplayComponents(new TextDisplayBuilder().setContent(headerBlock));
+      }
+      const [descriptionBlock, ...remainingBlocks] = bodyBlocks;
+      if (descriptionBlock) {
+        const descriptionSection = new SectionBuilder()
+          .addTextDisplayComponents(
+            new TextDisplayBuilder().setContent(descriptionBlock.content),
+          )
+          .setThumbnailAccessory(
+            new ThumbnailBuilder().setURL(thumbnailUrl),
+          );
+        container.addSectionComponents(descriptionSection);
+      }
+      remainingBlocks.forEach((block) => {
+        if (block.accessory) {
+          const section = new SectionBuilder()
+            .addTextDisplayComponents(new TextDisplayBuilder().setContent(block.content))
+            .setButtonAccessory(block.accessory);
+          container.addSectionComponents(section);
+        } else {
+          container.addTextDisplayComponents(new TextDisplayBuilder().setContent(block.content));
+        }
+      });
+    } else {
+      if (headerBlock.length > 0) {
+        container.addTextDisplayComponents(new TextDisplayBuilder().setContent(headerBlock));
+      }
+      bodyBlocks.forEach((block) => {
+        if (block.accessory) {
+          const section = new SectionBuilder()
+            .addTextDisplayComponents(new TextDisplayBuilder().setContent(block.content))
+            .setButtonAccessory(block.accessory);
+          container.addSectionComponents(section);
+        } else {
+          container.addTextDisplayComponents(new TextDisplayBuilder().setContent(block.content));
+        }
+      });
+    }
+
+    return {
+      components: [container],
+      files,
+      hasThread: Boolean(threadId),
+      featuredVideoUrl: game.featuredVideoUrl ?? null,
+      canMarkThumbnailBad: Boolean(game.artData) && !isThumbnailApproved,
+      isThumbnailBad,
+      isThumbnailApproved,
+      isReleased,
+    };
+  } catch (error: any) {
+    console.error("Failed to build game profile:", error);
+    return null;
+  }
+}
+
+export async function showGameProfile(
+  interaction: CommandInteraction | StringSelectMenuInteraction,
+  gameId: number,
+  includeActionsOverride?: boolean,
+  source?: GameSource,
+): Promise<void> {
+  const profile = await buildGameProfile(gameId, interaction, source);
+  if (!profile) {
+    await safeReply(interaction, {
+      content: `No game found with ID ${gameId}.`,
+      flags: MessageFlags.Ephemeral,
+    });
+    return;
+  }
+
+  const includeActions = includeActionsOverride ?? (
+    !("isMessageComponent" in interaction) ||
+    !interaction.isMessageComponent()
+  );
+  const components = [...profile.components];
+  if (includeActions) {
+    components.push(
+      ...buildGameProfileActionRow(
+        gameId,
+        profile.hasThread,
+        profile.featuredVideoUrl,
+        profile.canMarkThumbnailBad,
+        profile.isThumbnailBad,
+        profile.isThumbnailApproved,
+        profile.isReleased,
+      ),
+    );
+  }
+
+  await safeReply(interaction, {
+    embeds: [],
+    files: profile.files,
+    components,
+    flags: buildComponentsV2Flags(false),
+  });
+}
+
+export async function showGameProfileFromNomination(
+  interaction: StringSelectMenuInteraction,
+  gameId: number,
+): Promise<void> {
+  const profile = await buildGameProfile(gameId, interaction);
+  if (!profile) {
+    await safeReply(interaction, {
+      content: `No game found with ID ${gameId}.`,
+      flags: MessageFlags.Ephemeral,
+    });
+    return;
+  }
+  const components = [
+    ...profile.components,
+    ...buildGameProfileActionRow(
+      gameId,
+      profile.hasThread,
+      profile.featuredVideoUrl,
+      profile.canMarkThumbnailBad,
+      profile.isThumbnailBad,
+      profile.isThumbnailApproved,
+      profile.isReleased,
+    ),
+  ];
+  await safeReply(interaction, {
+    embeds: [],
+    files: profile.files,
+    components,
+    flags: buildComponentsV2Flags(true),
+  });
+}
+
+export async function buildGameProfileMessagePayload(
+  gameId: number,
+  options?: {
+    includeActions?: boolean;
+    guildId?: string;
+    prefaceText?: string;
+  },
+): Promise<{
+  components: Array<ContainerBuilder | ActionRowBuilder<ButtonBuilder>>;
+  files: AttachmentBuilder[];
+} | null> {
+  const includeActions = options?.includeActions ?? true;
+  const profile = await buildGameProfile(gameId, {
+    guildId: options?.guildId,
+    includeInlineButtons: includeActions,
+    prefaceText: options?.prefaceText,
+  });
+  if (!profile) {
+    return null;
+  }
+
+  const components = [...profile.components];
+  if (includeActions) {
+    components.push(
+      ...buildGameProfileActionRow(
+        gameId,
+        profile.hasThread,
+        profile.featuredVideoUrl,
+        profile.canMarkThumbnailBad,
+        profile.isThumbnailBad,
+        profile.isThumbnailApproved,
+        profile.isReleased,
+      ),
+    );
+  }
+
+  return {
+    components,
+    files: profile.files,
+  };
+}
+
+export async function refreshGameProfileMessage(
+  interaction: ButtonInteraction,
+  gameId: number,
+): Promise<void> {
+  const profile = await buildGameProfile(gameId, interaction);
+  if (!profile) return;
+  const actionRows = buildGameProfileActionRow(
+    gameId,
+    profile.hasThread,
+    profile.featuredVideoUrl,
+    profile.canMarkThumbnailBad,
+    profile.isThumbnailBad,
+    profile.isThumbnailApproved,
+    profile.isReleased,
+  );
+  const existingComponents = interaction.message?.components ?? [];
+  const searchRows = getSearchRowsFromComponents(existingComponents);
+  await interaction.editReply({
+    embeds: [],
+    files: profile.files,
+    components: [...profile.components, ...actionRows, ...searchRows],
+    flags: buildComponentsV2Flags(false),
+  }).catch(() => {});
+}
+
+export async function updateGameProfileMessageById(
+  interaction: ButtonInteraction | ModalSubmitInteraction,
+  channelId: string,
+  messageId: string,
+  gameId: number,
+): Promise<void> {
+  const channel = await interaction.client.channels.fetch(channelId).catch(() => null);
+  if (!channel || !channel.isTextBased() || !("messages" in channel)) {
+    return;
+  }
+  const message = await channel.messages.fetch(messageId).catch(() => null);
+  if (!message) return;
+  const profile = await buildGameProfile(gameId, interaction);
+  if (!profile) return;
+  const actionRows = buildGameProfileActionRow(
+    gameId,
+    profile.hasThread,
+    profile.featuredVideoUrl,
+    profile.canMarkThumbnailBad,
+    profile.isThumbnailBad,
+    profile.isThumbnailApproved,
+    profile.isReleased,
+  );
+  const searchRows = getSearchRowsFromComponents(message.components ?? []);
+  await message.edit({
+    embeds: [],
+    files: profile.files,
+    components: [...profile.components, ...actionRows, ...searchRows],
+  }).catch(() => {});
+}
